+++ src-rust/lib.rs
@@ -1,5 +1,3 @@
-use pyo3::create_exception;
-use pyo3::exceptions::PyException;
 use pyo3::prelude::*;
 use pyo3::Python;
 
@@ -33,6 +31,32 @@ impl CustomInstruction {
     }
 }
 
+/// Information about a custom classical function that should be defined in mathematical
+/// expressions.
+///
+/// The given `callable` must be a Python function that takes `n_params` floats, and returns a
+/// float.  The `name` is the identifier that refers to it in the OpenQASM 2 program.  This cannot
+/// clash with any defined gates.
+#[pyclass(text_signature = "(name, n_params, callable, /)")]
+#[derive(Clone)]
+pub struct CustomClassical {
+    pub name: String,
+    pub n_params: usize,
+    pub callable: PyObject,
+}
+
+#[pymethods]
+impl CustomClassical {
+    #[new]
+    fn __new__(name: String, n_params: usize, callable: PyObject) -> Self {
+        Self {
+            name,
+            n_params,
+            callable,
+        }
+    }
+}
+
 /// Create a bytecode iterable from a string containing an OpenQASM 2 program.  The iterable will
 /// lex and parse the source lazily; evaluating OpenQASM 2 statements as required, without loading
 /// the entire token and parse tree into memory at once.
@@ -42,12 +66,14 @@ fn bytecode_from_string(
     string: String,
     include_path: Vec<std::ffi::OsString>,
     custom_instructions: Vec<CustomInstruction>,
+    custom_classical: Vec<CustomClassical>,
     strict: bool,
 ) -> PyResult<bytecode::BytecodeIterator> {
     bytecode::BytecodeIterator::new(
         lex::TokenStream::from_string(string),
         include_path.iter().map(|x| x.into()).collect(),
         &custom_instructions,
+        &custom_classical,
         strict,
     )
 }
@@ -61,22 +87,17 @@ fn bytecode_from_file(
     path: std::ffi::OsString,
     include_path: Vec<std::ffi::OsString>,
     custom_instructions: Vec<CustomInstruction>,
+    custom_classical: Vec<CustomClassical>,
     strict: bool,
 ) -> PyResult<bytecode::BytecodeIterator> {
     bytecode::BytecodeIterator::new(
         lex::TokenStream::from_path(path)?,
         include_path.iter().map(|x| x.into()).collect(),
         &custom_instructions,
+        &custom_classical,
         strict,
     )
 }
-
-create_exception!(
-    core,
-    QASM2ParseError,
-    PyException,
-    "An error raised during parsing of OpenQASM 2 programs."
-);
 
 /// An interface to the Rust components of the parser stack, and the types it uses to represent the
 /// output.  The principal entry points for Python are :func:`bytecode_from_string` and
@@ -91,8 +112,10 @@ fn core(py: Python<'_>, module: &PyModule) -> PyResult<()> {
     module.add_class::<bytecode::ExprArgument>()?;
     module.add_class::<bytecode::ExprUnary>()?;
     module.add_class::<bytecode::ExprBinary>()?;
+    module.add_class::<bytecode::ExprCustom>()?;
     module.add_class::<CustomInstruction>()?;
-    module.add("QASM2ParseError", py.get_type::<QASM2ParseError>())?;
+    module.add_class::<CustomClassical>()?;
+    module.add("QASM2ParseError", py.get_type::<error::QASM2ParseError>())?;
     module.add_function(wrap_pyfunction!(bytecode_from_string, module)?)?;
     module.add_function(wrap_pyfunction!(bytecode_from_file, module)?)?;
     Ok(())